refactored the file structures that the temporal_network is only temporal network:)
